+++ src/main/java/team18/team18_be/recruitment/repository/RecruitmentRepository.java
@@ -1,0 +1,4 @@
+package team18.team18_be.recruitment.repository;
+
+public interface RecruitmentRepository {
+}
