+++ src/main/java/team18/team18_be/recruitment/service/RecruitmentService.java
@@ -1,0 +1,4 @@
+package team18.team18_be.recruitment.service;
+
+public class RecruitmentService {
+}
