--- src/main/java/team18/team18_be/repository/ResumeRepository.java
@@ -1,4 +1,0 @@
-package team18.team18_be.repository;
-
-public interface ResumeRepository {
-}
