+++ src/main/java/team18/team18_be/resume/dto/ResumeResponse.java
@@ -1,6 +1,6 @@
-package team18.team18_be.dto;
+package team18.team18_be.resume.dto;
 
-public record ResumeRequest(
+public record ResumeResponse(
         String resumeId,
         String applicantName,
         String address,
