+++ src/main/java/team18/team18_be/resume/repository/ResumeRepository.java
@@ -1,0 +1,4 @@
+package team18.team18_be.resume.repository;
+
+public interface ResumeRepository {
+}
