+++ src/main/java/team18/team18_be/resume/service/ResumeService.java
@@ -1,0 +1,4 @@
+package team18.team18_be.resume.service;
+
+public class ResumeService {
+}
