--- src/main/java/team18/team18_be/service/RecruitmentService.java
@@ -1,4 +1,0 @@
-package team18.team18_be.service;
-
-public class RecruitmentService {
-}
